leavemanagementsystem/requirements.txt
asgiref==3.2.3
astroid==2.2.5
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.5
Babel==2.6.0
bugsnag==3.5.2
certifi==2018.11.29
chardet==3.0.4
codecov==2.0.15
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
dataclasses==0.6
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-cors-headers==2.5.2
django-cors-middleware==1.3.1
django-environ==0.4.5
django-extensions==2.1.6
django-heroku==0.3.1
django-phonenumber-field==2.1.0
django-redis==4.10.0
django-rest-swagger==2.1.0
django-safedelete==0.5.2
django-stubs==0.10.2
django-versatileimagefield==1.10
djangorestframework==3.9.0
djangorestframework-stubs==0.4.2
entrypoints==0.3
environ==1.0
Faker==1.0.1
filelock==3.0.10
flake8==3.6.0
freezegun==0.3.14
furl==2.0.0
future==0.17.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==1.5.0
isort==4.3.16
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==5.0.0
mypy==0.670
mypy-extensions==0.4.1
numpy==1.16.4
oauthlib==3.0.1
openapi-codec==1.3.2
orderedmultidict==1.0
phonenumberslite==8.10.2
Pillow==5.4.1
pipenv==2018.11.26
pluggy==0.8.0
psycopg2==2.7.7
psycopg2-binary==2.7.6.1
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
PyJWT==1.7.1
pylint==2.3.1
pytest==4.0.2
pytest-cov==2.6.0
pytest-django==3.4.4
python-dateutil==2.8.1
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.7
redis==3.3.11
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
text-unidecode==1.2
toml==0.10.0
tox==3.6.1
typed-ast==1.3.1
typing-extensions==3.7.2
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.1.0
virtualenv-clone==0.5.1
WebOb==1.8.5
whitenoise==4.1.2
wrapt==1.11.1
yapf==0.26.0
zipp==2.1.0
